+++ .idea/workspace.xml
@@ -4,18 +4,8 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="c2025dea-9502-450e-9296-f399386a3f54" name="변경" comment="add: 예시 클래스 추가">
-      <change afterPath="$PROJECT_DIR$/.idea/uiDesigner.xml" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/MyMain.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/framework/ClassLoader.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/users/Bird.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/users/Cat.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/gradle.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/gradle.xml" afterDir="false" />
+    <list default="true" id="c2025dea-9502-450e-9296-f399386a3f54" name="변경" comment="feat: 패키지 로드 및 스캔">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build.gradle" beforeDir="false" afterPath="$PROJECT_DIR$/build.gradle" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Main.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/framework/LocationScanner.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/framework/LocationScanner.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/user/Dog.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/users/Dog.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -110,6 +100,9 @@
     &quot;accountId&quot;: &quot;d201b740-dfcd-46ff-8fa2-fa248dc0a2db&quot;
   }
 }</component>
+  <component name="PerforceDirect.Settings">
+    <option name="CHARSET" value="없음" />
+  </component>
   <component name="ProjectColorInfo">{
   &quot;associatedIndex&quot;: 4
 }</component>
@@ -125,7 +118,7 @@
   &quot;keyToString&quot;: {
     &quot;ASKED_ADD_EXTERNAL_FILES&quot;: &quot;true&quot;,
     &quot;Gradle.spring-core [:Main.main()].executor&quot;: &quot;Run&quot;,
-    &quot;Gradle.spring-core [:MyMain.main()].executor&quot;: &quot;Run&quot;,
+    &quot;Gradle.spring-core [:MyMain.main()].executor&quot;: &quot;Debug&quot;,
     &quot;Gradle.spring-core [build].executor&quot;: &quot;Run&quot;,
     &quot;Gradle.spring-core [clean].executor&quot;: &quot;Run&quot;,
     &quot;Gradle.spring-core [jar].executor&quot;: &quot;Run&quot;,
@@ -151,7 +144,7 @@
   </component>
   <component name="RunManager" selected="애플리케이션.MyMain">
     <configuration name="MyMain" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="MyMain" />
+      <option name="MAIN_CLASS_NAME" value="cns.MyMain" />
       <module name="spring-core.main" />
       <method v="2">
         <option name="Make" enabled="true" />
@@ -251,7 +244,7 @@
       <workItem from="1718112787557" duration="4464000" />
       <workItem from="1718541950754" duration="6255000" />
       <workItem from="1718719676150" duration="199000" />
-      <workItem from="1718789761400" duration="335000" />
+      <workItem from="1718789761400" duration="1204000" />
     </task>
     <task id="LOCAL-00001" summary="chore: 프로젝트 세팅">
       <option name="closed" value="true" />
@@ -277,7 +270,15 @@
       <option name="project" value="LOCAL" />
       <updated>1718115857340</updated>
     </task>
-    <option name="localTasksCounter" value="4" />
+    <task id="LOCAL-00004" summary="feat: 패키지 로드 및 스캔">
+      <option name="closed" value="true" />
+      <created>1718790134031</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1718790134032</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -303,19 +304,7 @@
       <map>
         <entry key="MAIN">
           <value>
-            <State>
-              <option name="FILTERS">
-                <map>
-                  <entry key="branch">
-                    <value>
-                      <list>
-                        <option value="origin/master" />
-                      </list>
-                    </value>
-                  </entry>
-                </map>
-              </option>
-            </State>
+            <State />
           </value>
         </entry>
       </map>
@@ -326,6 +315,7 @@
     <MESSAGE value="chore: 프로젝트 세팅" />
     <MESSAGE value="chore: README.md 추가" />
     <MESSAGE value="add: 예시 클래스 추가" />
-    <option name="LAST_COMMIT_MESSAGE" value="add: 예시 클래스 추가" />
+    <MESSAGE value="feat: 패키지 로드 및 스캔" />
+    <option name="LAST_COMMIT_MESSAGE" value="feat: 패키지 로드 및 스캔" />
   </component>
 </project>
