+++ src/main/java/cns/framework/ClassLoader.java
@@ -1,4 +1,4 @@
-package framework;
+package cns.framework;
 
 import java.lang.reflect.InvocationTargetException;
